feature/36 add Actions.tsx
